pyproject.toml
[project]
name = "sitcon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "paddlepaddle==3.0.0",
    "paddleocr>=3.0.0",
    "fastapi[standard]>=0.116.0",
    "google-genai>=1.25.0",
    "pillow>=11.3.0",
]
